Improves handling of unresponsive nodes
Before, jobs that were marked as running but were assigned to nodes that were stalled and unresponsive would be invisible to nodesforjob, because any nodes with "-" as one of their stats would be discarded.

Now, nodes will be marked as unresponsive or offline depending on whether only their current data is unavailable or all their data is unavailable, and this will be reported in place of load data.