Disables prompting for results pages in adname
Previously if you made a query that returned multiple results,
ldapsearch would page and prompt you to hit enter to see more results.

This didn't interact well with grep pipes, leaving the user thinking
that the command had stalled or maybe was waiting on stdin.

This change sets a Large page size and also disables prompting.