Adds email->username conversion to checkuser
Before, if you passed an email address to `checkuser`, it would `id` successfully because of the mysteries of Active Directory, but all the other checks would fail.

Now, it looks for an @ucl.ac.uk address in the parameter, and if it sees one, it tries to pass it to `id`, and tries to convert it to a username also using `id`. Then all other tests are run with the converted username.