Fixes to handle weird lfs error reporting
Turns out when `lfs quota` detects a problem, it prints out the numbers in brackets.

The bracketed numbers messed up the parsing of the usage figures (whatever good they are when lfs reports an error), and gave stack traces instead of user-intelligable messages.

This fixes the parsing so that if it detects bracketed numbers, it sets an error flag and then removes the brackets, and reports the error to the user.