lquota -u: fix double-counting
Once quotas are enabled, `lquota -u` would double-count space
when the lustre path is so long that the summary information in
`lfs quota -v` is printed on a separate line. (Happens on myriad's
/scratch.)

Fix by more rigourously filtering for `lfs quota -v` output lines
that come from the individual MDTs/OSTs. Also incorporate @ikirker's
suggestion to combine the original `grep` and `awk` commands (but
with a small fix to the `awk` regexp).

Tested on Myriad, Kathleen, Michael, Young, and Thomas.