:hammer: Updates scriptfor to use dated directories
Before, `scriptfor` just searched flat directories.

This assumed that one job ID uniquely matched one stored script.

Unfortunately the job IDs were reset to 1 so this is no longer the case, and job scripts are now archived in date-stamped YYYY-MM-DD subdirectories of the archive directory when they start. This means `scriptfor` now has to search for them instead of being able to specify the path just based on the job ID.

This change:
 - separates the treatment of the live directory, where scripts for running jobs are stored, and the archive directory.
 - means that it is no longer possible to specify an arbitrary number of search directories.
 - makes the script stop searching at the first match by default.
 - adds an option to keep searching for all matches.
 - adds option `-k` to specify the archive directory, using the old `-d` option to specify the active directory.
 - updates and expands the help info.